l3build: allow for missing support/ dirs, etc.
I've been experimenting with trying to minimise the
amount of additional material needed in a package directory
when using l3build. Before this commit, having 
"support/" was necessary even if empty.

This might not be the best place to make this check, though...

+++ l3build/l3build.lua
@@ -208,16 +208,18 @@ end
 
 -- Copy files 'quietly'
 function cp (glob, source, dest)
-  mkdir (dest)
-  for _,i in ipairs (filelist (source, glob)) do
-    local source = source .. "/" .. i
-    if os_windows then
-      os.execute (
-          "copy /y " .. unix_to_win (source) .. " "
-            .. unix_to_win (dest) .. " > nul"
-        )
-    else
-      os.execute ("cp -f " .. source .. " " .. dest)
+  if direxists(source) then
+    mkdir (dest)
+    for _,i in ipairs (filelist (source, glob)) do
+      local source = source .. "/" .. i
+      if os_windows then
+        os.execute (
+            "copy /y " .. unix_to_win (source) .. " "
+              .. unix_to_win (dest) .. " > nul"
+          )
+      else
+        os.execute ("cp -f " .. source .. " " .. dest)
+      end
     end
   end
 end
